pretty much done

+++ BLAS/BLAS3/main_dgemm.c
@@ -50,7 +50,7 @@ int main(int argc, char *argv[])
     free(C);
 
     printf("Runtime : %lf\n", time);
-    printf("Flops : %lf\n", ((float)M * N * K * 3 + M * N) / time / 1e9);
+    printf("Flops : %lf\n", ((double)M * N * K * 3 + M * N) / time / 1e9);
     printf("Bandwidth : %lf\n", (M * N + M * K + K * N) * sizeof(double) / time / 1e9);
 
     return 0;
