+++ BLAS/BLAS3/main_sgemm.c
@@ -49,7 +49,7 @@ int main(int argc, char *argv[])
     free(B);
 
     printf("Runtime : %lf\n", time);
-    printf("Flops : %lf\n", ((float)M * N * K * 3 + M * N) / time / 1e9);
+    printf("Flops : %lf\n", ((double)M * N * K * 3 + M * N) / time / 1e9);
     printf("Bandwidth : %lf\n", (M * N + N * K + M * K) * sizeof(float) / time / 1e9);
 
     return 0;
